+++ packages/wonder-blocks-tabs/src/components/navigation-tab-item.tsx
@@ -159,7 +159,7 @@ const styles = StyleSheet.create({
             border: "none",
             outline: "none",
             color: semanticColor.core.foreground.instructive.default,
-            backgroundColor: "transparent",
+            backgroundColor: semanticColor.core.transparent,
         },
         // NOTE: We use :not[aria-disabled] to avoid the hover styles to be
         // applied when the interactive element is disabled.