+++ packages/wonder-blocks-icon-button/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @khanacademy/wonder-blocks-icon-button
 
+## 11.3.1
+
+### Patch Changes
+
+- Updated dependencies [40cb70f]
+- Updated dependencies [40cb70f]
+- Updated dependencies [c97ece4]
+    - @khanacademy/wonder-blocks-core@12.4.4
+    - @khanacademy/wonder-blocks-clickable@8.2.0
+    - @khanacademy/wonder-blocks-icon@5.3.15
+    - @khanacademy/wonder-blocks-typography@4.3.5
+
 ## 11.3.0
 
 ### Minor Changes