+++ packages/wonder-blocks-typography/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @khanacademy/wonder-blocks-typography
 
+## 4.3.5
+
+### Patch Changes
+
+- Updated dependencies [40cb70f]
+- Updated dependencies [40cb70f]
+    - @khanacademy/wonder-blocks-core@12.4.4
+
 ## 4.3.4
 
 ### Patch Changes